Some modifications have been made due to the commonality of test codes.

+++ test/include/test/helper/constraint_metadata_helper.h
@@ -22,7 +22,7 @@ namespace manager::metadata::testing {
 
 class ConstraintMetadataHelper {
  public:
-  static int32_t get_record_count();
+  static int64_t get_record_count();
 };
 
 }  // namespace manager::metadata::testing
